requirements.txt
django
djangorestframework
psycopg2
python-dotenv
dj-rest-auth
django-allauth
djangorestframework-simplejwt
